# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Web Umkm",
        "description": "This For API System,And You Can Try It",
        "contact": {
            "email": "muhammadalghifari4321@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/jenistoko": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Looking Data JenisToko",
                "description": "Looking Data JenisToko",
                "operationId": "getjenisToko",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Insert Data JenisToko Admin",
                "description": "Insert Data JenisToko Admin",
                "operationId": "insertJenisToko",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "jenis_toko": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jenistoko/{id}": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Get By Id Data JenisToko Admin",
                "description": "Get By Id Data JenisToko Admin",
                "operationId": "getByIdjenisToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Get Data And Relation",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Update Data JenisToko Admin",
                "description": "Update Data JenisToko Admin",
                "operationId": "updateJenisToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "jenis_toko": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Delete Data JenisToko Admin",
                "description": "Delete Data JenisToko Admin",
                "operationId": "deleteJenisToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Delete",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    ]
}